Issue #32 Add a CFG (Control Flow Graph) builder
CFGBuilder.java:
Reworked the use of edgesToBuild instance member. It will now only be limited to the use of labels (to be implemented). Its content will also only be checked after visiting a statement, no longer before visiting it.
In newCFGBasicBlock(), now pass in the list of liveBlocks in order to link them to a newly created basicBlock.