Issue #69 Support building with Eclipse Tycho
Apparently, to test a bundle, the standard Tycho way  is to have a separate tests bundle.
I do not like very much this approach because it unnecessarily exposes packages (even with using x-friends) and it is slow to run the tests, but I have not found any other way to do it.

Added a plugin directory and a plugin.tests directory and build them as bundles with Tycho.
Moved unit tests to the plugin.tests directory. it does not compile because I am using TestNG. I need either to fix TestNG for Tycho or use JUnit.