Fixed issues with StringBuilderRefactoring
Ensured that when String appends start with 2 integers (or doubles, etc.) then a call to String.valueOf() is done instead.
Ensured that when String appends is only made of something else than a string, then a call to String.valueOf() is done instead.
Handled the following case: new String(Buffer|Builder)().toString()