Split the "samples" module from "plugin.tests" module
This allows adding TestNG refactorings with a testng dependency under maven.
Without this change, Tycho refuses to build the test plugin, because I cannot find an OSGi bundle which exposes the "org.testng" package which owns the Assert class.
Anyway, this will be helpful for refactoring APIs embedded in jars that are not OSGi bundles.