close #203 Vertical scroll bar on the Choose refactorings... dialog
The vertical scroll bar appears correctly on Linux Mint 17.3.
However it is not visible on Ubuntu 14, Ubuntu 16 and Windows 10.

It's not the bar that is hidden, it's the table viewer that overflows.
That's the reason why there is no bar: the viewer is as higher as the table.

This commit changes a property to force the viewer to suit the size of the window.
It also enlarges the filter text, to improve the feeling.