Disable JDK8 builds temporarily until release 1.1 is done
16 tests are failing under JDK 8, but not under JDK 7.
They are failing because some type bindings are unexpectedly `null` (the `int` type, for example).
I wonder if this is a bad combination of an old Eclipse target platform and a more recent JDK.
We will see if these test failures still exist when running against a more recent target platform.